+++ crates/ruma-client-api/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 Breaking changes:
 
+- Upgrade `js_option` to v0.2.0
+
+Improvements:
+
+- Stabilize support for the `use_state_after` query parameter and `State::After`
+  response property to `sync_events::v3`, according to Matrix 1.16.
+
+# 0.21.0
+
+Breaking changes:
+
 - Use `AuthType` for the `auth_type` of `get_uiaa_fallback_page`'s Request.
 - `get_supported_versions::Response::known_versions()` returns a
   `BTreeSet<MatrixVersion>` instead of a `DoubleEndedIterator`.